Release v2.1.0
Formalize interface definition language and data definition language.
The idia hers is: since SRPC use Python as IDL all content of <module_name>_interface.py and <module_name>.py must use type hint mandatorily.

+++ tests/integration/simple_project/calc/calc.py
@@ -11,5 +11,5 @@ class Calc(CalcInterface):
     def mult(self, a: int, b: int) -> int:
         return a * b
 
-    def div(self, a: int, b: int) -> int:
+    def div(self, a: int, b: int) -> float:
         return a / b
